Use conditional exports
