+++ packages/pwa-kit-runtime/src/utils/ssr-config.client.js
@@ -1,0 +1,11 @@
+/*
+ * Copyright (c) 2021, salesforce.com, inc.
+ * All rights reserved.
+ * SPDX-License-Identifier: BSD-3-Clause
+ * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
+ */
+
+export const getConfig = () => {
+    console.log('EXECUTING CLIENT IMPLEMENTATION')
+    return window.__CONFIG__
+}
