+++ packages/pwa/app/ssr.js
@@ -8,7 +8,8 @@
 
 const path = require('path')
 const {getRuntime, serveStaticFile} = require('pwa-kit-runtime/ssr/server/express')
-const {isRemote, loadConfig} = require('pwa-kit-runtime/utils/ssr-server')
+const {isRemote} = require('pwa-kit-runtime/utils/ssr-server')
+const {getConfig} = require('pwa-kit-runtime/utils/ssr-config')
 const helmet = require('helmet')
 
 const options = {
@@ -26,7 +27,7 @@ const options = {
     manifestPath: 'static/manifest.json',
 
     // This is the value of the 'mobify' object from package.json
-    mobify: loadConfig(),
+    mobify: getConfig(),
 
     // The port that the local dev server listens on
     port: 3000,
