+++ packages/pwa/app/utils/utils.js
@@ -5,7 +5,7 @@
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 
-import {getConfig} from 'pwa-kit-runtime/utils/ssr-shared'
+import {getConfig} from 'pwa-kit-runtime/utils/ssr-config'
 
 /**
  * Call requestIdleCallback in supported browsers.
